requirements.txt
dash
dash-bootstrap-components
plotly
yfinance
pandas
numpy
python-dotenv
anthropic
flask-caching
